+++ Changelog.md
@@ -11,6 +11,9 @@
 
 ## Gazebo 7.0.0 (2016-01-25)
 
+1. New world event: BeforePhysicsUpdate
+    * [Issue #1851](https://bitbucket.org/osrf/gazebo/issues/1851)
+
 1. Add FollowerPlugin
     * [Pull request #2085](https://bitbucket.org/osrf/gazebo/pull-request/2085)